chore: util cleanup
